more code style fixes

+++ src/utils.h
@@ -22,7 +22,7 @@ const char **
 strings_to_null_terminated(list_t *list);
 
 char
-get_size_unit(unsigned long kbytes, unsigned long *out_bytes);
+get_size_unit(unsigned long long kbytes, unsigned long *out_bytes);
 
 unsigned long
 parse_size_unit(const char *input);
